Bump and refactor
