+++ src/domain/core/components/header.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { Button } from "@/components/button";
+import { Button } from "@/domain/core/components/button";
 
 export const Header = () => {
   return (
