+++ src/domain/core/components/icon/icons/github.tsx
