+++ src/domain/core/components/icon/icons/mouse.tsx
