+++ src/domain/core/components/icon/index.ts
